+++ src/test.rs
@@ -12,8 +12,7 @@ fn construct_random() {
     ]) {
         eprintln!("RANDOM Testing n = {}", n);
         let keys = generate_keys(n);
-        let ptr_hash =
-            DefaultPtrHash::<RandomIntHash, _, _>::new(&keys, PtrHashParams::default_fast());
+        let ptr_hash = DefaultPtrHash::<IntHash, _, _>::new(&keys, PtrHashParams::default_fast());
         let mut done = bitvec![0; n];
         for key in keys {
             let idx = ptr_hash.index(&key);
@@ -31,7 +30,7 @@ fn test_1e9() {
     let n = 1_000_000_000;
     eprintln!("RANDOM Testing n = {}", n);
     let keys = generate_keys(n);
-    let ptr_hash = DefaultPtrHash::<RandomIntHash, _, _>::new(&keys, PtrHashParams::default_fast());
+    let ptr_hash = DefaultPtrHash::<IntHash, _, _>::new(&keys, PtrHashParams::default_fast());
     let mut done = bitvec![0; n];
     for key in keys {
         let idx = ptr_hash.index(&key);
@@ -49,13 +48,13 @@ fn int_hash_speed() {
 
     let start = std::time::Instant::now();
     for k in &keys {
-        black_box(RandomIntHash::hash(k, seed));
+        black_box(IntHash::hash(k, seed));
     }
     eprintln!("Time {:?}", start.elapsed());
 
     let start = std::time::Instant::now();
     for k in &keys {
-        black_box(IntHash::hash(k, seed));
+        black_box(StrongerIntHash::hash(k, seed));
     }
     eprintln!("Time {:?}", start.elapsed());
 
@@ -84,7 +83,7 @@ fn construct_multiples() {
             eprintln!("MULTIPLES OF {m} Testing n = {}", n);
             let keys = (0..n as u64).map(|i| m * i).collect::<Vec<_>>();
             let ptr_hash =
-                DefaultPtrHash::<IntHash, _, _>::new(&keys, PtrHashParams::default_fast());
+                DefaultPtrHash::<StrongerIntHash, _, _>::new(&keys, PtrHashParams::default_fast());
             let mut done = bitvec![0; n];
             for key in keys {
                 let idx = ptr_hash.index(&key);
@@ -129,7 +128,7 @@ fn in_memory_sharding() {
     let n = 1 << 25;
     let range = 0..n as u64;
     let keys = range.clone().into_par_iter();
-    let ptr_hash = <PtrHash<_, _, Vec<u32>, IntHash, _>>::new_from_par_iter(
+    let ptr_hash = <PtrHash<_, _, Vec<u32>, StrongerIntHash, _>>::new_from_par_iter(
         n,
         keys.clone(),
         PtrHashParams {
@@ -152,7 +151,7 @@ fn on_disk_sharding() {
     let n = 1 << 25;
     let range = 0..n as u64;
     let keys = range.clone().into_par_iter();
-    let ptr_hash = <PtrHash<_, _, Vec<u32>, IntHash, _>>::new_from_par_iter(
+    let ptr_hash = <PtrHash<_, _, Vec<u32>, StrongerIntHash, _>>::new_from_par_iter(
         n,
         keys.clone(),
         PtrHashParams {
@@ -178,7 +177,7 @@ fn many_keys_memory() {
     let n_query = 1 << 27;
     let range = 0..n as u64;
     let keys = range.clone().into_par_iter();
-    let ptr_hash = <PtrHash<_, _, Vec<u32>, IntHash, _>>::new_from_par_iter(
+    let ptr_hash = <PtrHash<_, _, Vec<u32>, StrongerIntHash, _>>::new_from_par_iter(
         n,
         keys.clone(),
         PtrHashParams {
@@ -208,7 +207,7 @@ fn many_keys_disk() {
     let n_query = 1 << 27;
     let range = 0..n as u64;
     let keys = range.clone().into_par_iter();
-    let ptr_hash = <PtrHash<_, _, Vec<u32>, IntHash, _>>::new_from_par_iter(
+    let ptr_hash = <PtrHash<_, _, Vec<u32>, StrongerIntHash, _>>::new_from_par_iter(
         n,
         keys.clone(),
         PtrHashParams {